+++ Package.swift
@@ -18,7 +18,10 @@ let package = Package(
             url: "https://github.com/RevenueCat/purchases-ios-spm.git",
             exact: "5.67.0"
         ),
-        .package(path: "../Kiki_mackit")
+        .package(
+            url: "https://github.com/Feng6611/Kiki_mackit.git",
+            exact: "0.8.2"
+        )
     ],
     targets: [
         .target(name: "KikiCommerceCore"),